Modal now appears when head is clicked
