+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/useradd/workspace/Transitionr/transitionr/src/index.js":"1","/home/useradd/workspace/Transitionr/transitionr/src/reportWebVitals.js":"2","/home/useradd/workspace/Transitionr/transitionr/src/App.js":"3","/home/useradd/workspace/Transitionr/transitionr/src/MainPage.js":"4","/home/useradd/workspace/Transitionr/transitionr/src/HeadMenu.js":"5"},{"size":504,"mtime":1611605450805,"results":"6","hashOfConfig":"7"},{"size":362,"mtime":1611596450073,"results":"8","hashOfConfig":"7"},{"size":252,"mtime":1611608488719,"results":"9","hashOfConfig":"7"},{"size":536,"mtime":1611608637054,"results":"10","hashOfConfig":"7"},{"size":1649,"mtime":1611608421006,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"damc6s",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"19","messages":"20","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/useradd/workspace/Transitionr/transitionr/src/index.js",[],["23","24"],"/home/useradd/workspace/Transitionr/transitionr/src/reportWebVitals.js",[],"/home/useradd/workspace/Transitionr/transitionr/src/App.js",["25"],"/home/useradd/workspace/Transitionr/transitionr/src/MainPage.js",["26","27","28","29"],"/home/useradd/workspace/Transitionr/transitionr/src/HeadMenu.js",[],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":1,"column":10,"nodeType":"36","messageId":"37","endLine":1,"endColumn":15},{"ruleId":"34","severity":1,"message":"38","line":3,"column":10,"nodeType":"36","messageId":"37","endLine":3,"endColumn":18},{"ruleId":"34","severity":1,"message":"39","line":6,"column":10,"nodeType":"36","messageId":"37","endLine":6,"endColumn":18},{"ruleId":"34","severity":1,"message":"40","line":13,"column":13,"nodeType":"36","messageId":"37","endLine":13,"endColumn":17},{"ruleId":"41","severity":1,"message":"42","line":20,"column":13,"nodeType":"43","endLine":20,"endColumn":51},"no-native-reassign",["44"],"no-negated-in-lhs",["45"],"no-unused-vars","'Route' is defined but never used.","Identifier","unusedVar","'useState' is defined but never used.","'HeadMenu' is defined but never used.","'coor' is assigned a value but never used.","jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
+[{"/Users/ardisred/Workspace/congruence/src/index.js":"1","/Users/ardisred/Workspace/congruence/src/App.js":"2","/Users/ardisred/Workspace/congruence/src/reportWebVitals.js":"3","/Users/ardisred/Workspace/congruence/src/MainPage.js":"4"},{"size":504,"mtime":1611609361404,"results":"5","hashOfConfig":"6"},{"size":252,"mtime":1611609361400,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":1611609361405,"results":"8","hashOfConfig":"6"},{"size":1910,"mtime":1611612440211,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"fzwdfm",{"filePath":"13","messages":"14","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"15","usedDeprecatedRules":"12"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"20"},"/Users/ardisred/Workspace/congruence/src/index.js",[],["21","22"],"/Users/ardisred/Workspace/congruence/src/App.js",["23"],"import { Route, BrowserRouter } from \"react-router-dom\"\nimport React from \"react\"\nimport { MainPage } from './MainPage'\n\n\nexport const App = (props) => {\n\n  return <>\n      <BrowserRouter>\n\n        <MainPage {...props} />\n      </BrowserRouter>\n  </>\n}","/Users/ardisred/Workspace/congruence/src/reportWebVitals.js",[],"/Users/ardisred/Workspace/congruence/src/MainPage.js",["24"],"\nimport React from 'react'\nimport { useState } from 'react'\nimport './App.css';\nimport Body from './illust57-4505.jpg'\nimport ListGroup from 'react-bootstrap/ListGroup';\nimport Button from 'react-bootstrap/Button';\nimport Modal from 'react-bootstrap/Modal'\n\nexport function MainPage(props) {\n\n    const [show, setShow] = useState(false);\n    const handleClose = () => setShow(false);\n    const handleShow = () => setShow(true);\n\n    const handleControlledInputChange = (e) => {\n        var x = e.clientX;\n        var y = e.clientY;\n        var coor = \"X coords: \" + x + \", Y coords: \" + y;\n        if (x >= 401 && x <= 482 && y >= 62 && y <= 181){\n            handleShow();\n        }\n        console.log(coor)\n    }\n\n    return <>\n    <div>\n            <Modal show={show} onHide={handleClose} animation={false}>\n                <Modal.Body>   <div className=\"headMenu\">\n                    <ListGroup>\n                        <ListGroup.Item>Cras justo odio</ListGroup.Item>\n                        <ListGroup.Item>Dapibus ac facilisis in</ListGroup.Item>\n                        <ListGroup.Item>Morbi leo risus</ListGroup.Item>\n                        <ListGroup.Item>Porta ac consectetur ac</ListGroup.Item>\n                        <ListGroup.Item>Vestibulum at eros</ListGroup.Item>\n                    </ListGroup>\n                </div>\n                </Modal.Body>\n                <Modal.Footer>\n                    <Button variant=\"secondary\" onClick={handleClose}>\n                        Close\n            </Button>\n                    <Button variant=\"primary\" onClick={handleClose}>\n                        Save Changes\n            </Button>\n                </Modal.Footer>\n            </Modal></div>\n        <div onClick={handleControlledInputChange} className=\"fullScreen\">\n            <img onClick={handleControlledInputChange} className=\"bodyImage\" src={Body}></img>\n        </div>\n    </>\n}\n\n\n",{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","severity":1,"message":"30","line":1,"column":10,"nodeType":"31","messageId":"32","endLine":1,"endColumn":15},{"ruleId":"33","severity":1,"message":"34","line":49,"column":13,"nodeType":"35","endLine":49,"endColumn":89},"no-native-reassign",["36"],"no-negated-in-lhs",["37"],"no-unused-vars","'Route' is defined but never used.","Identifier","unusedVar","jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
