--- src/HeadMenu.css
@@ -1,3 +1,0 @@
-.headMenu {
-    border: solid 1px red
-}
